requirements.txt
requests==2.31.0
Pillow==11.0.0
SpeechRecognition==3.10.0
googletrans==4.0.0rc1
httpx==0.13.3
pyaudio==0.2.14
